+++ newsletter/package.json
@@ -1,19 +1,19 @@
 {
   "name": "newsletter",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "private": true,
   "scripts": {
     "deploy": "wrangler deploy",
     "dev": "wrangler dev",
     "start": "wrangler dev"
   },
+  "dependencies": {
+    "hono": "^3.7.6",
+    "zod": "^3.22.4"
+  },
   "devDependencies": {
     "@cloudflare/workers-types": "^4.20230419.0",
     "typescript": "^5.0.4",
     "wrangler": "^3.0.0"
-  },
-  "dependencies": {
-    "hono": "^3.7.6",
-    "zod": "^3.22.4"
   }
 }
