+++ newsletter/src/utils/response.ts
@@ -1,0 +1,1 @@
+export const getResponse = (message: string) => ({ message })
