+++ newsletter/src/utils/types.ts
@@ -1,0 +1,6 @@
+import { Env } from 'hono'
+import { Env as CloudflareEnv } from './constants'
+
+export interface HonoEnv extends Env {
+    Bindings: CloudflareEnv
+}
